backend/requirements.txt
alembic==1.12.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.1.0
exceptiongroup==1.3.1
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0

